requirements.txt
Flask>=3.0
qrcode[pil]>=7.4
Pillow>=10.0
Flask>=3.0
qrcode[pil]>=7.4
Pillow>=10.0
gunicorn
